pyproject.toml
[tool.poetry]
name = "oursns"
version = "0.1.0"
description = ""
authors = ["25th-Night <brown_eyed87@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "3.2.13"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.0"
Pillow = "^9.2.0"
django-cors-headers = "^3.13.0"
django-filter = "^22.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
